pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gif-maker"
version = "1.0.1"
description = "Professional Animated GIF Creation Tool"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.4.0,<12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
gif-maker = "gif_maker:main"

[tool.setuptools.packages.find]
include = ["gif_maker*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
